requirements.txt
discord-anti-spam==1.2.3
nextcord